made changes

+++ src/main/java/searchengine/repository/LemmaRepository.java
@@ -4,10 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import searchengine.model.Lemma;
 import searchengine.model.Site;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface LemmaRepository extends JpaRepository<Lemma, Integer> {
+public interface LemmaRepository extends JpaRepository<Lemma, Long> {
     Optional<Lemma> findByLemmaAndSite(String lemma, Site site);
     Optional<Lemma> findByLemma (String lemma);
     int countBySite(Site site);
