+++ src/main/java/searchengine/repository/SiteRepository.java
@@ -6,7 +6,7 @@ import searchengine.model.Site;
 
 
 @Repository
-public interface SiteRepository extends JpaRepository <Site, Integer> {
+public interface SiteRepository extends JpaRepository <Site, Long> {
 
     Site findByUrl(String url);
